requirements.txt
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
kornia==0.6.8
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
wincertstore==0.2

setproctitle~=1.3.2